duui-Stance/requirements.txt
tqdm==4.66.1
transformers==4.38.2
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
sentencepiece==0.2.0
autoawq==0.2.4
openai==1.14.3
scipy==1.10.1
dkpro-cassis==0.9.1
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic-settings==2.0.2